+++ src/app/shared/components/side-nav/side-nav.component.ts
@@ -99,6 +99,10 @@ export class SideNavComponent implements AfterViewInit {
     window.open('https://www.youtube.com/playlist?list=PLSMRtuVN409fB35RLljjg3jNkVJbLIP1u', '_blank');
   }
 
+  public OpenGlossary() {
+    this.dialog.OpenGlossaryDialog();
+  }
+
   private hovered: string;
   public OnMouseEnter(route: string) { this.hovered = route; }
   public OnMouseLeave() { this.hovered = ''; }