+++ src/app/shared/shared.module.ts
@@ -74,6 +74,7 @@ import { TagsComponent } from './components/tags/tags.component';
 import { CveSearchComponent } from './components/cve-search/cve-search.component';
 import { CveEntryComponent } from './components/cve-entry/cve-entry.component';
 import { ImageViewComponent } from './components/image-view/image-view.component';
+import { GlossaryComponent } from './components/glossary/glossary.component';
 
 @NgModule({
   declarations: [
@@ -81,7 +82,7 @@ import { ImageViewComponent } from './components/image-view/image-view.component
     StatusBarComponent, SideNavComponent, SaveDialogComponent, PasswordDialogComponent, MessagesDialogComponent, 
     TwoOptionsDialogComponent, CweEntryComponent, CapecEntryComponent, NavTreeComponent, CvssEntryComponent, OwaspRREntryComponent,
     ProgressTrackerComponent, ModelInfoComponent, NotesComponent, 
-    LocalDatePipe, LocalDateTimePipe, RenameDialogComponent, TagsComponent, CveSearchComponent, CveEntryComponent, ImageViewComponent
+    LocalDatePipe, LocalDateTimePipe, RenameDialogComponent, TagsComponent, CveSearchComponent, CveEntryComponent, ImageViewComponent, GlossaryComponent
   ],
   imports: [
     CommonModule,